Use `all` pattern instead of `any` for checking
When checking something for compliance make the code return `false` in
case of an error, instead of returning `true` in case of a success. This
code pattern is less error prone since the function only reaches the
bottom (and returns `true`) in the case every previous condition is met.
The other way may lead to false `true` returns.